requirement.txt
Django==2.2.7
djangorestframework==3.10.3
pocketsphinx==0.1.15
PyAudio==0.2.11
pytz==2019.3
SpeechRecognition==3.8.1
sqlparse==0.3.0
